Update BalanceByYears and BalanceByAge implementations
Replaced `int` with `byte` for `Age` and `Years` to improve datatype precision and consistency. Added a new service method for fetching balance reports grouped by years. Adjusted related endpoints, mappers, and CSV output to align with these changes.

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Response/YearEnd/Frozen/BalanceByAgeDetail.cs
@@ -1,16 +1,11 @@
 ﻿namespace Demoulas.ProfitSharing.Common.Contracts.Response.YearEnd.Frozen;
+
 public sealed record BalanceByAgeDetail : BalanceByDetailBase
 {
-    public required int Age { get; init; }
-   
+    public required byte Age { get; init; }
+
     public static BalanceByAgeDetail ResponseExample()
     {
-        return new BalanceByAgeDetail
-        {
-            Age = 32,
-            CurrentBalance = (decimal)159_451.46,
-            EmployeeCount = 7,
-            BeneficiaryCount = 3
-        };
+        return new BalanceByAgeDetail { Age = 32, CurrentBalance = (decimal)159_451.46, EmployeeCount = 7, BeneficiaryCount = 3 };
     }
 }
